Chatbot/requirements.txt
langchain 
langchain_openai
langchain_groq 
langchain_community 
langgraph 
langsmith 
typing-extensions
python_dotenv
